+++ src/GreenWallet/WalletConfig.h
@@ -65,5 +65,5 @@ namespace WalletConfig
     /* If a mixin of zero is disabled, at what height was it disabled? E.g.
        fork height, or 0, if never allowed. This is ignored if a mixin of
        zero is allowed */
-    const uint64_t mixinZeroDisabledHeight = 0;
+    const uint32_t mixinZeroDisabledHeight = 0;
 }
